Refactor Design Smell : UMLFactory class refactored into ClassDiagram and PackageDiagram visitor implementations.
